+++ inst/tst_pkgs/dummy_pkg_errors/R/test.R
@@ -1,5 +1,5 @@
 #' @keywords internal
 tstCharLimitLintr <- function() {
-  x   <- 1
+  x   <- 1 # workaround for GDR-1234
   print("This file is simply a test to see whether or not the lintr functions appropriately detect lines that are too long.")
 }